Upgrade to VSSDK.BuildTools 17.14.1043-preview2
The original reason to upgrade to this is carries along updates
to the Fast Up to Date Check support, which means Roslyn can delete
our own copy of the targets.

The upgrade caused one break: there's newer support in the VS SDK now
where they look for ForceIncludeInVSIX metadata on PackageReferences
as a way to bypass the suppression logic that exists to ensure certain
references don't end up in the final VSIX. We had our own custom logic
where you'd set just IncludeInVSIX="true" as metadata, which would
produce items under teh covers with ForceIncludeInVSIX=true. There's
arguably a bug in the targets that they should be respecting our
underlying metadata rather than looking back to the PackageReference
metadata, but it still seems reasonable for us just to adopt the same
metadata name as the VS SDK. At this point it's not clear to me if
we even need our custom logic anymore or if it's now obsolete.

# Conflicts:
#	eng/Versions.props
#	src/VisualStudio/Setup.Dependencies/Roslyn.VisualStudio.Setup.Dependencies.csproj