Further reduce thread switches during package load. (#77602)
Further reduce thread switches during package load.

The Rolsyn package load sequence is quite confusing, and currently does quite a few thread switches. Trying re-order statements is difficult, and fraught with danger if one of the calls is entered or returns on an unexpected thread. For example, one component created during load, the ColorSchemeApplier, needs to be initialized during this sequence, but does up to three thread switches during it's initialization. Every time work is added to the package load sequence, we run a pretty high risk of adding new thread switches during this critical period.

Instead, this change allows packages to add work to main thread or background thread work queues. Items in each of these queues are batched together and can use a single thread switch for the entire batch. These queues can be added to during processing, allowing packages to "discover" new work that needs to be processed as they proceed.