Support ILogger.BeginScope() for LSP logging
This surrounds the processing of each LSP request with a call to
ILogger.BeginScope() with the name of the method that's being invoked.
This is prefixed to the message we send back to the LSP client, so
it's much easier to understand the source of various pieces of logging.

ILspLogger.LogStartContext/LogEndContext are deprecated here, as
setting scopes plus an explicit start/end as appropriate makes things
work much better when we're handling requests in parallel.