Another attempt to move MiscellaneousFilesWorkspace to a bg thread
This previously failed with integration test failures. I attribute this to previously trying to move the VisualStudioMetadataReferenceManager when that very much wants to be on the main thread. Instead, just ensure that the code that creates the VisualStudioMetadataReferenceManager  is on the main thread when that value is needed (it already was).

1) Move MiscellaneousFilesWorkspace to bg thread MEF construction
2) Move HostServicesProvider to location of VisualStudioMefHostServices and use in a couple places (this prevents MiscellaneousFilesWorkspace from needing to MEF import VisualStudioWorkspace in it's ctor)