Move the construction cost of SettingsEditorFactory to when it's used (#77754)
SettingsEditorFactory was noted in a previous PR as having it's construction and registration separated across threads, having it's construction occur on a background thread due to it's cost. However, all of the data MEF imported by the ctor isn't used until it's CreateEditorInstance method is called, and can thus be deferred until then.

Note that I also moved this to be a singleton exposed through a static method rather than MEF. I originally coded this to just change all the ctor params to be lazy, but this still showed up as 4 ms of cost (which was an improvement over the 22 ms of cost without using lazys) over multiple runs of release bits. Instead, just putting it into a static had a 0 ms cost over multiple runs.

Also:
1) Got rid of an Assumes.Present as MEF will have already thrown before that point
2) Got rid of some Contract.Throws that were checking things that the containing method would never do.