Make Razor redirect code more robust
In an attempt to normalize the Razor experience the Workspaces layer
will redirect Razor analyzer references to the ones checked into
Visual Studio. This detection had a couple of corner cases that it
didn't handle:

1. When customers used Razor from anywhere other than the .NET SDK. This
   would impact customers that check in artifacts as part of their
   build.
2. In a future where Razor changes the set of DLLs that are registered
   with `/analyzer` and customers use an older SDK, this logic would
   attempt to remap DLLs that did not exist. That is because it just
   assumes that DLLs coming from the SDK exist in VS.

This attempts to make the process more robust by identifying an analyzer
as being part of the Razor closure if it exists in the same directory as
the Razor compiler. Then for all such DLLs it will just return the set
of registered Razor analyzers.

This does have one remaining corner case where a customer registers in an
analyzer in a directory with a file named MS.CA.Razor.Compiler.dll but
that file is not registered as an analyzer. That seems a bit out of
bounds as they're shipping MS.CA.Razor.Compiler in a state that is not
supported. Happy to hear if anyone has ideas on how to solve this
though.