Move test helpers to NetLatest
This is the first of a series of PRs to move our tests to use
`NetLatest` by default. Using `NetLatest` means that our tests
use `net472` reference assemblies when running on .NET Framework
and latest .NET Core reference assemblies when running on .NET
Core. This is advantageous in a few areas:

1. This is much closer to how our customers operate
2. It mean that as we move the _latest_ definition in our unit test
   infra our tests move forward automatically

This change moves a number of our core helpers from `Standard`, which is
`netstandard2.0` on .NET Core, to `NetLatest`. This should significantly
expand our coverage against modern .NET reference assemblies

This first commit updates our test infrastructure so that tests can set
a default`TargetFramework` that all `CreateCompilation` use. This makes
it easy for an individual test class to focus on say `netstandard2.0`, `net472`,
etc ...

The end goal is for the default experience to be `NetLatest` but that
requires changes to a large number of tests. To let me break these up
into individual PRs the default is currently per test assembly.
Subsequent PRs will move each test assembly to the default and fix up
test classes along the way.

Once everything is moved to `NetLatest` as a default, I will start going
back to high impact test classes that opted into `Standard` and move
them to `NetLatest`. That will result in much smaller and reviewable PRs.

infra