Move MiscellaneousFilesWorkspace construction to background thread (#78051)
This PR restores behavior introduced in #77768 that were reverted in PR #77983.
The revert occurred as it was flagged for causing ddrit regressions in some of the WinformsVS64.Designer tests. I was unable to reproduce this scenario locally,
but the RPS tests are able to reliably hit this issue.

Digging into it a bit indicated that creating the OpenTextBufferProvider on a bg thread somehow a thread affinity that caused
microsoft.visualstudio.shell.design.ni!DocData.CreateNativeDocData's call to msenv!CEditorManager::RegisterInvisibleEditor's to require an RPC main->bg thread switch.
This breaks the assumptions of RegisterInvisibleEditor and weird things start happening. Jason helped track down the likely culprit, and logged
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/2441480 to track that work.

This PR works around that issue by obtaining the RDT while on the main thread. I've logged #78050 to track the Roslyn cleanup
necessary once obtaining the rdt can be done from a background thread.

a) Restores changes from PR #77768 that were reverted in PR #77983
b) Ensure OpenTextBufferProvider._runningDocumentTable is initialized on the main thread.
c) Added comments about https://devdiv.visualstudio.com/DevDiv/_workitems/edit/2441480 and created a tracking bug for us to clean this up once it's fixed (#78050)