WIP: Start moving workspace events to fire on background threads
This addresses a major portion of https://github.com/dotnet/roslyn/issues/78091

There are still some WorkspaceChanged events that haven't been switched over. This is big enough without doing all those and I'd like to get a ci run on this to see how broken tests are.

Additionally, this didn't tackle adding the WorkspaceChangedEventOptions, that can be added later if we determine it would provide value.