Ensure FindSourceDefinitionAsync works in frozen compilation cases
FindSourceDefinitionAsync was assuming that in the cross-language
case, if you have a symbol from another project, that the compilation
must always be available for that project -- for example, that calling
TryGetCompilation on the project must always work. That's not the
case if we're not generating skeleton references all the time. The
fix to make that method work reliably again is simply to call
GetCompilationAsync(). This results in a cascade of making more things
async though: although our public entrypoint for
FindSourceDefinitionAsync was Task-returning, it wasn't async and
called into a non-async helper. So this removes those internal helpers
and makes everything async again.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/2443981
Fixes https://developercommunity.visualstudio.com/t/Go-to-definition-stopped-working-after-u/10796494