Allocate slightly less in ImmutableArrayExtensions.ConditionallyDeOrder
This method currently allocates two arrays (once during the ToArray call, the other in the AsImmutable call). This is generally not a problem as this is debug only code, but this method is called very frequently, so I'll change the line to make this method only allocate the array once.