Do not use nint / nuint in IL baselines
This removes the use of `nint / nuint` in our default IL display
baselines. Using them caused issues for our tests that run on both .NET
Core and .NET Framework TFMs as it caused `IntPtr / UIntPtr` to display
different in each case. This is an attempt to normalize our baselines so
it's more resilient to tests like this.

In general, our IL baselines should reflect metadata names, not C# type
names, to further this goal. Did consider removing `UseSpecialNames`
from `ILVisualizationFormat` but that was very disruptive.

closes #78240