Small allocation improvement in AbstractRecommendationService.GetRecommendedSymbolsInContext
Reduce allocations in AbstractRecommendationService.GetRecommendedSymbolsInContext by realizing two immutable arrays only once instead of twice. Previously, FilterToVisibleAndBrowsableSymbols returned an IA and then we filtered it and realized it again via a call to WhereAsArray. Instead, modify FilterToVisibleAndBrowsableSymbols to take in a filter that it can use before realizing the array.

This change removes about 50 MB of allocations (~1%) in the C# completion section of the roslyn editing speedometer test

See PR for images representing the allocation patterns before/after