Convert SyntaxNode.GetText users to instead retrieve the source text from the document (#78259)
Remove some of the SyntaxNode.GetText callers that were calling this on a root node, and instead switch them over to use Document.GetValueTextAsync instead. Each SyntaxNode.GetText call builds up a new SourceText from scratch. Instead, if Document.GetValueTextAsync is used, there is the potential for sharing the sourcetext with other document consumers.

This shows up as about a 50MB allocation improvement in VS under AbstractMemberInsertingCompletionProvider.RemoveDestinationNodeAsync in the override completion commit scenario in the C# editing speedometer.