Reduce allocations in NamespaceSymbol.GetExtensionContainers (#78243)
Reduce allocations from executing NamespaceSymbol.GetExtensionContainers. This method accounts for 48% of allocations in our CodeAnalysis process during completion in our C# editing speedometer test. The changes in this PR reduce that to about 4.5%

The change here is to add an override of NamespaceSymbol.GetExtensionContainers to MergedNamespaceSymbol. Specifically, the base class implementation of GetExtensionContainers calls GetTypeMembersUnordered, which is inefficient when in a MergedNamespaceSymbol instance as it does a Linq OfType call and allocates an IA from that. Instead, the new override can just enumerate over the result of GetMembersUnordered directly, and do the check before calling into shared code with the base implementation.

See PR for images representing before/after allocation patterns