Avoid creating identity to obtain public key (#78255)
Reduce allocations in AssemblySymbol.GivesAccessTo due to getting the assemblyWantingAccess's Identity.

During the completion scenarios in the C# speedometer tests, about 3% of allocations in the VS process are used obtaining this identity just to get the PublicKey. However, in the case where assemblyWantingAccess is an AssemblySymbol, there is a way to obtain this without needing to obtain the full Identity.

See PR for images outlining before / after allocation patterns.