For IsAtDotDotToken, ensure the current token is a DotToken before parsing the next token to see if it's also a DotToken.
I see LanguageParser.IsAtDotToken accounting for about 1 second of CPU time (0.5%) in the C# editing speedometer test. Local testing indicates about 85% of calls into this method will be filtered by this check.

Going to initiate a speedometer run via a tes insertion and will update with the results when available.