Implement SeparatedSyntaxList.Count(Func<T, bool>)
CSharpAttributeData.IsTargetEarlyAttribute was calling the Linq version of this list, and thus boxing the SeparatedSyntaxList enumerator, showing up as 6.5 MB (0.1%) of allocations in our OOP process during solution load.

Not a huge win, but it's nice to add to the pit of success for using this type.

Modifications to AttributeData just started with changing IsTargetEarlyAttribute to demonstrate the Count usage. I then looked to see what other Linq usage was in the file, and it made sense to change those too.