Implement SeparatedSyntaxList.Count(Func<T, bool>) (#78348)
CSharpAttributeData.IsTargetEarlyAttribute was calling the Linq version of this list, and thus boxing the SeparatedSyntaxList enumerator, showing up as 5.9 MB (0.1%) of allocations in our OOP process during solution load.

Not a huge win, but it's nice to add to the pit of success for using this type.

Modifications to CSharpAttributeData started with a small cleanup to IsTargetEarlyAttribute to demonstrate the Count() usage. I then looked to see what other Linq usage was in the file, and it made sense to change those too.

See PR for image with defails about allocations being addressed.