Update `RuntimeHelpers.Await` rules (#77957)
These rules now clarify the extensibility point we're looking to establish with the runtime for `Await` helpers. Generally speaking, these are just ordinary methods that a user is free to call, so there's no issue with us using them for subtypes of `Task`, since the user could write that themselves and the runtime will have to handle it. Also removed is `HoistedLocal`, as it will no longer be used by the runtime.