Make compiler detection more resilient
This changes the way our build task finds the csc to launch to longer
depend on the `CompilerType` task argument. At the core for every
deployment kind of the task there is only one possible compiler that can
be loaded. Hence change our logic to better assert which deployment we
are in and locate the compiler based on that.