Change `++`/`--` predefined operator names for instance operators (#78434)
See https://github.com/dotnet/csharplang/pull/9356

With this change instance `++`/`--` operators get different predefined method names (as opposed to static `++`/`--` operators). Compiler assigns method names to operators pretty early and it does that based on syntax. Given that, we need a simple and reliable heuristic to do that for operator declarations and for CRef references. This change implements the following heuristic for `++`/`--` operators – if there is an empty argument list in the syntax, the construct is treated as instance operator and gets instance operator names, otherwise it is treated as static operator and gets static operator names. We cannot use `static` modifier for the heuristic because it is not allowed in CRef.