Reduce allocations under ContextMutableIntervalTree
There are some closure allocations unnecessarilly happening during ContextMutableIntervalTree construction. Simply changed a couple methods to static and added a parameter for the extra bit of data they need to get rid of those allocations.

This is a very small improvement, only about 0.1% (6 MB) of allocations during typing in the razor speedometer test.