Take reachability into account when rewriting `await`s in `finally` blocks
Our current approach to rewriting `await` expressions in `finally` blocks is technically unsound: the rewrite can result in a new exit block from the original method being created. As a simple example, here is an original method, and what it gets rewritten to by `AsyncExceptionHandlerRewriter` (slightly simplified):

```cs
static async Task G()
{
    int x = 0;

    try
    {
        x = await F();
    }
    finally
    {
        x += await F();
    }
}
```

```cs
async Task<int> M()
{
    int x = 0;
    object ex = null;
    int retTaken = 0;
    int retTemp = 0;
    try
    {
        x = await F();
        goto @finally;
    }
    catch (object o)
    {
        ex = o;
        goto @finally;
    }

@finally:
    x += await F();
    if (ex != null)
    {
        if (ex is Exception e) ExceptionDispatchInfo.Capture(e).Throw();
        throw ex;
    }

    switch(retTaken)
    {
        case 2:
            return retTemp;
        default:
            break; // This is the problem
    }
}
```

While we can statically walk the method and prove via flow analysis that the `default` case of the switch isn't reachable, the generalized version of this problem is the halting problem. To date, this has never been a problem because the code that the async rewriter rewrites is then wrapped rewritten again into the async state machine, and the missing return becomes completely unobservable. However, for runtime async, we're going to be emitting this general structure, essentially unchanged: that `break;`, then, becomes a branch to an invalid IL location and the method ends without a `ret`. To solve this, we need to take reachability analysis into consideration when rewriting `try/finally`s. If the end of the `finally` is reachable, meaning the method continues beyond the end of the `try/finally`, then the `break;` is necessary, as that's the path by which the rest of the method will be reached. However, in the case where the end of the `try/finally` is not reachable, we need to unconditionally return, rather than letting it fall through. The other case would be if there was no returns, just unconditional throws in the body of the method. In this case, we want to insert a `throw null;` to the end of the switch to ensure that the exit remains covered. The problematic code above becomes this instead:

```cs
async Task<int> M()
{
    int x = 0;
    object ex = null;
    int retTaken = 0;
    int retTemp = 0;
    try
    {
        x = await F();
        goto @finally;
    }
    catch (object o)
    {
        ex = o;
        goto @finally;
    }

@finally:
    x += await F();
    if (ex != null)
    {
        if (ex is Exception e) ExceptionDispatchInfo.Capture(e).Throw();
        throw ex;
    }

    return retTemp;
}
```