Fix SubText ctor parameter verification.
This ctor currently throws given a zero length span at the end (and only the end) of the subtext. This evidently became far more prevalent with this (fairly) recent change to optimize newline information in our sourcetext classes. (https://github.com/dotnet/roslyn/pull/74728)

It doesn't make to allow zero length spans at all locations but at the end of the subtext, so the fix is just to allow construction in that case too.

Big props to Manish Vasani for providing a very actionable repro for this.

Fixes https://github.com/dotnet/roslyn/issues/78985.

Note there is a potentially related issue outlined in https://github.com/dotnet/roslyn/issues/76225 that looks attributable to the PR above too. It possible that this addresses that issue too, but I'm not completely convinced.