Add nullability annotations to ProjectExternalErrorReporter
There was one bug in the old code: there was a null check against
bstrFileName, which, if that was indeed null, would have thrown an
exception later when that null string was passed to a constructor.
I can find no evidence in telemetry or old bugs that indeed that
can be null, so I removed the check.