Delete unnecessary GetServiceAsync calls
It's not entirely clear to me why these were added; the intent may have
been to create them asynchronously before we had a later GetService()
call, but these services are likely to be created when we're loaded
(like IVsShell), so these won't do anything. Some of the other ones
we've now all moved to GetServiceAsync() at the actual use sites so
there's no reason to have that either.

Closes https://github.com/dotnet/roslyn/issues/29598