Add a workaround to keep unit tests working
Right now it appears we have a bug that when that update runs,
it updates the Solution.WorkspaceVersion which breaks the ability
for TryApplyChanges to work later. CodeModel doesn't deal with that case
so leaving this in causes CodeModel tests to get flaky.

Furthermore, the WhenActivated() calls can't be unsubscribed, and will
leak for later tests, which can cause our tests to OOM in CI.