Do not do full cleanup when we are formatting a new document
The CodeAction.CleanupDocumentAsync was added in
https://github.com/dotnet/roslyn/pull/59091 to ensure the document
had elastic trivia formatted before we applied the result to the
file. It appears the intent was to do just the trivia formatting, but
the CleanupDocumentAsync does a lot of expensive operations around
simplification and adding imports, which aren't needed in this case.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1540532