Get rid of an unnecessary image list load
We were fetching an image list on the UI thread during the registration
of our library manager for the Object Browser. This is surprisingly
expensive (50ms), and it turns out seems to be entirely unnecessary.
Returning zero means the default list is used, which works just fine.
In fact -- better than fine! This also fixes some fuzzy icons in
High DPI settings.