Reduce allocations in CSharpVirtualCharService.TryConvertStringToVirtualChars
This method is accounting for 1.4% of allocations in the OOP process in a completion scenario in the RazorEditing.CompletionInCohosting speedometer test I am looking at.

Instead of doing all the work and hitting the pools, we can just handle the normal case of there not being any surrogates or escape characters and just create a VirtualCharSequence wrapper around the existing string.

Locally, I see about 75% of strings that come through hit this optimization.

Going to do a test insertion with this change to verify speedometer numbers before elevating out of draft status.