Reduce some UI thread blocking
It looks like in various refactorings we accidentally had a
ConfigureAwait(false) appear in this method which could cause us to jump
to a background thread when we don't need to do that. I'm not sure why
we didn't just use CA(true) only here, since that seems to do what we
want without any extra trickery.