Reduce alllocations in AnalyzerExecutor.GetAddDiagnostic
Will create a speedometer run to verify this removes these allocations.

This method accounts for 2.4% of total allocations in the Roslyn ServiceHub process for the CompletionInCohosting speedometer test

There is already an existing mechanism used for pooling similar actions used in the ExecuteSemanticModelActions and ExecuteSyntaxTreeActions methods, so this just follows that pattern and moves the invocation code for compilation and symbol diagnostics into AnalyzerDiagnosticReporter.