Fix file-based programs getting stuck in the host workspace
If a file is loaded that can be treated as a file-based program,
and then a real project is loaded that contains that file, we should
unload the file-based program since we're treating that as a
miscellaneous file. But we had a bug here: the code in
LspWorkspaceManager that loops through workspaces assumes that the
miscellaneous files workspace would be found last, so if we find a
document in any other workspace first, we're good to use it. This
assumption was broken once we started putting file-based programs
in the host workspace, since we might find both the file-based
program and the 'real' project in the same workspace.

The fix is straightforward: if when we're finding a file context we amd
we end up with a miscellaneous file, check if we have any other files
too, and if so, prefer those. This will then allow us to unload the
miscellaneous file as normal.

Fixes https://github.com/dotnet/vscode-csharp/issues/8471