*** WIP: Reduce allocations in DeclataionComputer
This is a draft until I get back numbers from a speedometer run.

CSharpDeclarationComputer.ComputeDeclarations shows up as 1.0% of Roslyn OOP allocations during a completion scenario in the RazorEditingTests.CompletionInCohosting speedometer test. The majority of the allocations under this method appear to be linq related, so switching to the model where we use a pooled collection to collect the items rather than linq enumerables.